Renames "hex" cookie obfuscation to "hash". Adjusts tests.
Previously JAR used an oxy fork and custom cookie obfuscator suite
because oxy didn't support it natively. After years the PR was
partially accepted but the hex obfuscator wasn't, and someone
else's hash implementation was. This is just catchup as hex
has been hash since the re-adoption of mainline oxy.