Removes DocsOut subsystem.
DocsOut was an invaluable subsystem during development to ensure the
running server was properly reading the configs. This has been
unmaintained for years, leading to massive docrot. Moving to consts
for the config keys has vastly reduced fatfinger config errors, and
vociferous runtime debugging has replaced most of the value.