Refactors the ForbiddenPaths check in BuildPath.
The check previously errored out if the Path matched a
ForbiddenPath, but now it removes the matching ForbiddenPath
from THIS PATHs list (e.g. not globally).

Also of note, this commit introduces the passing of a reference
to the Path to BuildPath, so we can inspect the Path to see
changes to the ForbiddenPath.