Brings forth a yaegi-based plugin subsystem, disabled by default.
I originally wrote a version of this in 2020, but it became clear
that Yaegi was not made for long-running processes (no GC) nor
particularly performant (2 orders of magnitude slower on identical
code), so not of interest for this project. Fast forward to 2023
and it "was much better" so I updated this, but the caveats were
the same. Fast forward to now, and someone showed interest in
migrating if they didn't have to recompile to add middleware. Fine.
Please read and understand docs/Plugins.md before enabling.