Makes sweeping changes as to where certain code is located.
I got carried away. This should have been smaller grafts. Anywhere
in the core processes that AWS-specific "stuff" is called, it has been
abstracted, with several new (and undocumented) hooks to accomodate
the portability, the new aws.go file to house the bootstap, etc.

This has been tested with all of the AWS-specific files behind a
build tag, and everything builds and works just fine. Of note, TUS
will require a minute of rework when the time comes, as I backed
myself into a corner there. No worries. A fully de-AWSd jard
is under 20MB, and the memory profile is a fraction. Crazy how much
plumbing goes into that.