Removes pointless bound to healthCheck results chan.
Actually hit this bound during a contrived massive deployment. It
was originally envisioned as a gate to keep the healthchecks from
impacting overall performance, but the opposite is true: hitting
the bound and pausing the sending goros builds more resource plaque
than the result handler spinning at fullspeed processing results.