BREAKS workers as you knew them. Implements simplified worker pool.
The preceding line is alarmist, nothing is actually broken.
The workers/ subpackage has a new NewSimpleWorkerPool creator
that returns a version of WorkerPool that ignores all of the
configuration tunables that existed for the "old" WorkerPool,
in lieu of a simple goro loop that creates Workers and gives
them Work to do. It's a drop-in replacement.

JAR implements this "simple" WorkerPool immediately, and ignores
previous config directives. That may be a surprise. All code
that used workers/ will continue to work against this and subsequent
v1 releases unchanged and as-expected. In v2, NewWorkerPool will
return a "simple" pool, and most of the plumbing will be ripped out.
If you feel your JAR needs the not-simple WorkerPool, let's talk.